+++ ServerHost.sln
@@ -5,11 +5,13 @@ VisualStudioVersion = 15.0.27004.2006
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServerHost", "ServerHost\ServerHost.csproj", "{98D61E3E-76A7-4664-BAF7-FBBF41A798D9}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "Tests\Tests.csproj", "{DB6BD075-C56F-4335-83FD-DD4CEF140FED}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{A20AA77B-112B-4786-BDC1-7ADC8CE32C40}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestServer", "TestServer\TestServer.csproj", "{3C8985C0-B0EC-4EB6-8EA3-DCB7D770F3FC}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests.Net46", "Tests\Tests.Net46\Tests.Net46.csproj", "{DB6BD075-C56F-4335-83FD-DD4CEF140FED}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{64DBC33D-89B0-4F25-8E6D-3A4FAA8F4AE0}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestServer", "Tests\TestServer\TestServer.csproj", "{3C8985C0-B0EC-4EB6-8EA3-DCB7D770F3FC}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Build", "Build", "{64DBC33D-89B0-4F25-8E6D-3A4FAA8F4AE0}"
 	ProjectSection(SolutionItems) = preProject
 		.gitignore = .gitignore
 		.travis.yml = .travis.yml
@@ -18,10 +20,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		Build.sh = Build.sh
 		Clean.cmd = Clean.cmd
 		documentation\DocFormatter.xsl = documentation\DocFormatter.xsl
-		init.cmd = init.cmd
-		init.ps1 = init.ps1
 		nuget-restore.cmd = nuget-restore.cmd
-		nuget.config = nuget.config
 		ServerHost.nuspec = ServerHost.nuspec
 		Test.cmd = Test.cmd
 		Test.runsettings = Test.runsettings
@@ -30,6 +29,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{BAC26871-088E-4528-B17C-C847091AFC13}"
 	ProjectSection(SolutionItems) = preProject
+		nuget.config = nuget.config
 		.nuget\packages.config = .nuget\packages.config
 	EndProjectSection
 EndProject
@@ -54,6 +54,10 @@ Global
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{DB6BD075-C56F-4335-83FD-DD4CEF140FED} = {A20AA77B-112B-4786-BDC1-7ADC8CE32C40}
+		{3C8985C0-B0EC-4EB6-8EA3-DCB7D770F3FC} = {A20AA77B-112B-4786-BDC1-7ADC8CE32C40}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {20B9C9F1-4BE3-470E-81E8-2BB73541ED6A}
