+++ Tests/Tests.Net46/packages.config
